chainlit_app/requirements.txt
chainlit>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
azure-identity>=1.15.0
weaviate-client>=4.0.0
uvicorn>=0.23.0
gunicorn>=21.2.0
duckdb>=0.9.0  # Using duckdb instead of asyncpg
psycopg2-binary>=2.9.5
requests>=2.31.0
Pillow>=10.0.0
